% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glove_helpers.R
\name{closest_to}
\alias{closest_to}
\title{closest_to}
\usage{
closest_to(word_vec, word_vectors, top_n = 10)
}
\arguments{
\item{word_vectors}{all vectors from glove model}

\item{word}{word vector to return}
}
\description{
function that will determine words that are close to specific word
}
