% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glove_helpers.R
\name{glove_closest_to}
\alias{glove_closest_to}
\title{glove_closest_to}
\usage{
glove_closest_to(word_vectors, word_vec, top_n = 10)
}
\arguments{
\item{word_vec}{word vector to calculate on}

\item{top_n}{top n to return}
}
\description{
return the word associations closest to provided word
}
