% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glove.R
\name{plot_glove}
\alias{plot_glove}
\title{plot_glove}
\usage{
plot_glove(word_sim, txt_size = 15, bg = "transparent")
}
\arguments{
\item{word_sim}{output from glove estimate}

\item{txt_size}{max text size in wordcloud}

\item{bg}{background of ggplot}
}
\value{
tibble
}
\description{
brings back the text of journals that contain a certain word
}
