% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glove_helpers.R
\name{similarities}
\alias{similarities}
\title{similarities}
\usage{
similarities(word_vectors, word_vec)
}
\arguments{
\item{word_vec}{word vector to calculate on}
}
\description{
return the word similarities in terms of provided words
}
