% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glove.R
\name{model_glove}
\alias{model_glove}
\title{model_glove}
\usage{
model_glove(
  txt,
  ngrams = 1,
  term_count_min = 5,
  skip_grams_window = 10L,
  runs = 10,
  embedding_dim = 200,
  co_oc_max,
  cores = 8,
  verbose = TRUE
)
}
\arguments{
\item{txt}{character vector of the text to analyse}
}
\value{
tibble
}
\description{
runs a GloVe model on the text within the articles or provided texts
}
